+++ test/util.py
@@ -8,7 +8,7 @@ wordlist_funcs = [('wordlist_init', c_void_p, [c_char_p]),
                   ('wordlist_free', None, [c_void_p])]
 
 mnemonic_funcs = [('mnemonic_from_bytes', c_char_p, [c_void_p, c_void_p, c_ulong]),
-                  ('mnemonic_to_bytes', c_int, [c_void_p, c_char_p, c_void_p, c_ulong])]
+                  ('mnemonic_to_bytes', c_ulong, [c_void_p, c_char_p, c_void_p, c_ulong])]
 
 bip39_funcs = [('bip39_get_wordlist', c_void_p, [c_char_p]),
                ('bip39_mnemonic_from_bytes', c_char_p, [c_void_p, c_void_p, c_ulong])]
